Add files via upload
Using Python to construct a Pairs Trading Strategy. This model pulls the current list of S&P 500 companies as listed on wikipedia and allows the user to easily filter them by Sector and Industry which then tests for correlation and cointegration between pairs of stocks, normalizes their standard deviations and constructs buy and sell signals. A lot of inspiration came from https://kidquant.com/project/pairs-trading-strategies-in-python/